Update test to pass with 3.18 library - this no longer seems to throw an IOException, but instead returns FIBITMAP.Zero.
I considered updating the managed code to throw an IOException on an empty stream for compatibility - however, I think this behaviour is technically more correct since an empty stream is not really an exceptional circumstance. The new behaviour is to return an empty result given an empty input, which doesn't feel unintuitive.