Constrain the managed <-> unmanaged buffer copies to fit within a buffer no larger than the maximum that can be rented from the shared ArrayPool.
This prevents the ArrayPool from allocating a large array which then does not get returned to the pool (.Return() just ignores it).

This was resulting in unnecessary allocations. There are now no allocations during stream read/write, and therefore no garbage collections either.

Performance has increased as a result.

Addresses #22.