Fix issue with writing JPEG streams using the SpanStreamIO class - in this case, size = 1 and count = 4096 (writing BMPs, count = 1 and size = length of bitmap).
The code didn't account for count properly. I've decided not to do a record-by-record copy, since each record is only 1 byte and so this results in a lot of small copies. Instead, I'm just copying the whole buffer in an all-or-nothing manner.

Also added a new WriteStreamJpeg() test which would have caught this earlier.

Addresses #22.